+++ packages/vue-location/src/utils.ts
@@ -1,4 +1,4 @@
-import type { _RouteLocationBase } from 'vue-router';
+import type { _RouteLocationBase, Router, RouteLocationRaw } from 'vue-router';
 
 function clone<T>(source: T): T {
   return JSON.parse(JSON.stringify(source));
@@ -17,4 +17,20 @@ export function pickShallowRoute(
     redirectedFrom: route.redirectedFrom,
     meta: clone(route.meta),
   };
+}
+
+const normalizePath = (path: string) =>
+  path.endsWith('/') ? path : `${path}/`;
+
+export function locationIsMatched(router: Router, target: RouteLocationRaw) {
+  const currentRoute = router.currentRoute.value;
+  const _target = router.resolve(target);
+  const { path, query, hash } = currentRoute;
+  if (normalizePath(_target.path) !== normalizePath(path)) return false;
+  const queryEntries = Object.entries(query);
+  if (queryEntries.some(([key, value]) => _target.query[key] !== value)) {
+    return false;
+  }
+  if (hash && _target.hash !== hash) return false;
+  return true;
 }